Moves exceptions into a sub-namespace
Exceptions were almost fully half of the classes/interfaces/traits in
the package, and deserved their own namespace.
