+++ doc/book/exception.md
@@ -7,7 +7,7 @@ details.
 To facilitate this, we provide an interface, `ProblemDetailsException`:
 
 ```php
-namespace ProblemDetails;
+namespace ProblemDetails\Exception;
 
 use JsonSerializable;
 
@@ -50,8 +50,8 @@ transaction problem details, you might do so as follows:
 
 ```php
 use DomainException;
-use ProblemDetails\CommonProblemDetailsException;
-use ProblemDetails\ProblemDetailsException;
+use ProblemDetails\Exception\CommonProblemDetailsException;
+use ProblemDetails\Exception\ProblemDetailsException;
 
 class TransactionException extends DomainException implements ProblemDetailsException
 {
