+++ src/Exception/InvalidResponseBodyException.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace ProblemDetails;
+namespace ProblemDetails\Exception;
 
 use RuntimeException;
