+++ test/ProblemDetailsMiddlewareTest.php
@@ -5,9 +5,10 @@ namespace ProblemDetailsTest;
 use ErrorException;
 use Interop\Http\ServerMiddleware\DelegateInterface;
 use PHPUnit\Framework\TestCase;
-use ProblemDetails\MissingResponseException;
+use ProblemDetails\Exception\MissingResponseException;
 use ProblemDetails\ProblemDetailsMiddleware;
 use ProblemDetails\ProblemDetailsResponseFactory;
+use ProblemDetailsTest\TestAsset;
 use Prophecy\Argument;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
